About to change edit year (in setup) to 2 text fields rather than buttons for the NCEA level

+++ NCEACredits/Profile.h
@@ -21,7 +21,7 @@
 
 @property NSUInteger *currentYear;
 - (Year *)getYearObjectForYearDate:(NSUInteger)date;
-- (int)getYearCurrentlyInUseOtherwiseCurrentDateYear;
+- (NSUInteger)getYearCurrentlyInUseOtherwiseCurrentDateYear;
 
 @property YearCollection *yearCollection;
 - (NSArray *)getYearsAsTableDatasForSetup;
